requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# AI and Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
feedparser>=6.0.0
anthropic>=0.7.8

# Data handling
pandas>=2.0.0
openpyxl>=3.1.2

# Web dashboard
flask>=3.0.0

# Optional dependencies for enhanced features
psutil>=5.9.0  # For system stats
geopy>=2.3.0   # For geographic calculations

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0